requirements.txt
Flask==2.2.2
Flask-HTTPAuth
flask-marshmallow
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
marshmallow-sqlalchemy
loguru~=0.6.0
pymssql
python-dotenv~=0.21.0
requests~=2.28.1
pandas~=1.5.2
aniso8601
asgiref
nest_asyncio
websockets~=10.4
SQLAlchemy~=1.4.45
protobuf~=3.17.3
numpy
python-dateutil~=2.8.2
alembic~=1.8.1
gunicorn
marshmallow~=3.19.0
Werkzeug~=2.2.2
xmltodict~=0.13.0
APScheduler~=3.10.1
WTForms~=3.0.1
Flask-WTF
mysql
pyodbc
pymongo